pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-provenance"
version = "0.1.0"
description = "Git-native AI code provenance and metadata tracking system"
authors = [
    {name = "AI Provenance Contributors"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["git", "ai", "provenance", "metadata", "tracking", "audit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "click>=8.0.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ai-prov = "ai_provenance.cli.main:cli"

[project.urls]
Homepage = "https://github.com/ai-provenance/ai-provenance"
Documentation = "https://ai-provenance.readthedocs.io"
Repository = "https://github.com/ai-provenance/ai-provenance"
Issues = "https://github.com/ai-provenance/ai-provenance/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ai_provenance --cov-report=term-missing"
